+++ src/components/Fieldset/fieldsetChangelogData.ts
@@ -10,6 +10,13 @@ import { ChangelogData } from "../../utils/ComponentChangelogTable";
 
 export const changelogData: ChangelogData[] = [
   {
+    date: "2025-03-20",
+    version: "3.5.5",
+    type: "Update",
+    affects: ["Styles"],
+    notes: ['Changes the font weight for the `(required)` label to `"light"`.'],
+  },
+  {
     date: "2024-09-19",
     version: "3.3.2",
     type: "Update",