+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 ### Adds
 
 - Adds the `useResponsiveSpacing` hook.
+- Adds docs for `Responsive Grid` Style Guide.
 
 ### Updates
 
@@ -18,6 +19,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 - Updates the values for the breakpoints and the associated docs.
 - Updates the `viewport` preview options in Storybook to align with the Reservoir breakpoints.
 - Updates the Storybook font styles.
+- Updates the docs for the `Template` component.
 
 ## Prerelease